Added CoinlistItem and moved dirs
