+++ nexo-ta/src/components/Platform/Coinlist/index.module.css
